Change LogLocalIdx to uint16_t

+++ src/idx.h
@@ -12,8 +12,8 @@ using VirtualBlockIdx = uint32_t;
 // local index within a block; this can be -1 to indicate an error
 using BitmapLocalIdx = int16_t;
 using TxLocalIdx = int16_t;
-// Note: LogLocalIdx will persist and we need range [0, 256)
-using LogLocalIdx = uint8_t;
+// Note: LogLocalIdx will persist and the valid range is [0, 255]
+using LogLocalIdx = uint16_t;
 
 // identifier of bitmap blocks; checkout BitmapBlock's doc to see more
 using BitmapBlockId = uint32_t;
